:recycle: refactor(notification): change repository and service code
