+++ backend/streetdrop-batch/src/main/java/com/depromeet/domains/notification/dto/request/AllPushRequestDto.java
@@ -13,4 +13,8 @@ public class AllPushRequestDto {
 
     @NotNull
     private String content;
+
+    private String path;
+
+    private String pid;
 }
