+++ backend/streetdrop-batch/src/main/java/com/depromeet/domains/notification/service/TokenService.java
@@ -1,10 +1,13 @@
 package com.depromeet.domains.notification.service;
 
+import com.depromeet.common.error.code.GlobalErrorCode;
 import com.depromeet.domains.notification.dto.request.TokenRequestDto;
 import com.depromeet.common.error.code.TokenErrorCode;
 import com.depromeet.common.error.exceptions.NotFoundException;
-import com.depromeet.domains.notification.repository.UserDeviceRepository;
-import com.depromeet.domains.notification.domain.UserDevice;
+import com.depromeet.domains.user.repository.UserDeviceRepository;
+import com.depromeet.domains.user.repository.UserRepository;
+import com.depromeet.user.User;
+import com.depromeet.user.UserDevice;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -14,17 +17,21 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class TokenService {
 
+    private final UserRepository userRepository;
     private final UserDeviceRepository userDeviceRepository;
 
     @Transactional
     public void saveToken(TokenRequestDto tokenRequestDto) {
+        User user = userRepository.findById(tokenRequestDto.getUserId())
+                .orElseThrow(() -> new NotFoundException(GlobalErrorCode.USER_NOT_FOUND));
         Optional<UserDevice> userDevice = userDeviceRepository.findByUserId(tokenRequestDto.getUserId());
         if (userDevice.isPresent()) {
             UserDevice updatedUserDevice = userDevice.get().updateDeviceToken(tokenRequestDto.getToken());
             userDeviceRepository.save(updatedUserDevice);
-        } else {
+        }
+        else {
             UserDevice createdUserDevice = UserDevice.builder()
-                    .userId(tokenRequestDto.getUserId())
+                    .user(user)
                     .deviceToken(tokenRequestDto.getToken())
                     .build();
             userDeviceRepository.save(createdUserDevice);
@@ -37,4 +44,5 @@ public class TokenService {
                 .orElseThrow(() -> new NotFoundException(TokenErrorCode.TOKEN_NOT_FOUND));
         userDeviceRepository.delete(userDevice);
     }
+
 }
