+++ backend/streetdrop-batch/src/main/java/com/depromeet/domains/notification/service/TopicService.java
@@ -6,8 +6,8 @@ import com.depromeet.common.error.code.TokenErrorCode;
 import com.depromeet.common.error.exceptions.ExternalServerException;
 import com.depromeet.common.error.exceptions.NotFoundException;
 import com.depromeet.domains.notification.fcm.FcmService;
-import com.depromeet.domains.notification.repository.UserDeviceRepository;
-import com.depromeet.domains.notification.domain.UserDevice;
+import com.depromeet.domains.user.repository.UserDeviceRepository;
+import com.depromeet.user.UserDevice;
 import com.google.firebase.messaging.FirebaseMessagingException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
